#48: Move fragments into own folder
