+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/SectionsPagerAdapter.java
@@ -12,9 +12,9 @@ import android.content.Context;
 import android.support.v13.app.FragmentPagerAdapter;
 import android.util.SparseArray;
 import android.view.ViewGroup;
-import ca.ualberta.cs.shinyexpensetracker.activities.ClaimSummaryFragment;
-import ca.ualberta.cs.shinyexpensetracker.activities.DestinationListFragment;
-import ca.ualberta.cs.shinyexpensetracker.activities.ExpenseItemListFragment;
+import ca.ualberta.cs.shinyexpensetracker.fragments.ClaimSummaryFragment;
+import ca.ualberta.cs.shinyexpensetracker.fragments.DestinationListFragment;
+import ca.ualberta.cs.shinyexpensetracker.fragments.ExpenseItemListFragment;
 
 /**
  * A {@link FragmentPagerAdapter} that returns a fragment corresponding to
