+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/fragments/ClaimSummaryFragment.java
@@ -1,4 +1,4 @@
-package ca.ualberta.cs.shinyexpensetracker.activities;
+package ca.ualberta.cs.shinyexpensetracker.fragments;
 
 import java.io.IOException;
